% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_aliases.R, R/project.init.R
\name{dirdata}
\alias{dirdata}
\alias{dirdata}
\title{Data Dir
Helper wrapper to get data for this project.}
\usage{
dirdata(...)

dirdata(...)
}
\description{
Data Dir
Helper wrapper to get data for this project.

Data Dir
Helper wrapper to get data for this project.
}
