% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirAliases.R
\name{dirOut}
\alias{dirOut}
\title{Direct a path into the project output directory}
\usage{
dirOut(...)
}
\description{
Pass a relative file path, and this will prepend the global project results
directory to it. Use it to stick output easily directly into the results
directory, instead of relative to the local directory.
This allows you to keep a working directory that's relative to your code,
but put the results somewhere else (such as a shared results space).
}
