% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_aliases.R, R/project.init.R
\name{dirraw}
\alias{dirraw}
\alias{dirraw}
\title{Raw Data Dir
Helper wrapper to get data for this project.}
\usage{
dirraw(...)

dirraw(...)
}
\description{
Raw Data Dir
Helper wrapper to get data for this project.

Raw Data Dir
Helper wrapper to get data for this project.
}
