% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_aliases.R, R/project.init.R
\name{dirres}
\alias{dirres}
\alias{dirres}
\title{Resource Dir
Helper wrapper to get data for this project.}
\usage{
dirres(...)

dirres(...)
}
\description{
Resource Dir
Helper wrapper to get data for this project.

Resource Dir
Helper wrapper to get filesystem data for this project.
}
