% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_aliases.R, R/project.init.R
\name{dirweb}
\alias{dirweb}
\alias{dirweb}
\title{Web Dir
Helper wrapper to get data for this project.}
\usage{
dirweb(...)

dirweb(...)
}
\description{
Web Dir
Helper wrapper to get data for this project.

Web directory
Helper wrapper to get web data for this project.
}
