% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirAliases.R
\name{dirWrap}
\alias{dirWrap}
\title{Generic function to prepend an environment variable directory
to your relative filepath.}
\usage{
dirWrap(var, ..., sub = NULL)
}
\description{
Generic function to prepend an environment variable directory
to your relative filepath.
}

