% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.niceGetEnv}
\alias{.niceGetEnv}
\title{Make suggestion about configuring an environment variable.}
\usage{
.niceGetEnv(varname)
}
\arguments{
\item{varname}{Name of environment variable to suggest setting.}
}
\value{
Message about benefit of setting the given environment variable.
}
\description{
Make suggestion about configuring an environment variable.
}
