% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.selectPath}
\alias{.selectPath}
\title{Returns a full path}
\usage{
.selectPath(target, parent, default)
}
\arguments{
\item{target}{The path to check for seeming absolute-ness.}

\item{parent}{Name of the environment variable with parent folder candidate.}

\item{default}{Default to use if target is null.}
}
\value{
\code{target} if it's already absolute, result of \code{when_null()} 
  if \code{target} is null, or joined version of parent candidate stored in 
  \code{env_var} and (relative) \code{target}.
}
\description{
Returns a full path
}
\seealso{
Other path operations: \code{\link{.isAbsolute}}
}
\concept{path operations}
