% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{firstFile}
\alias{firstFile}
\title{Locator of first file from given sequence that exists.}
\usage{
firstFile(files, modify = identity)
}
\arguments{
\item{files:}{The sequence file names or paths to consider.}

\item{modify:}{How to modify each file before checking existence.}
}
\value{
(Absolute) path to the first file that exists.
        \code{NULL} if there isn't one.
}
\description{
\code{firstFile} examines the filesystem and selects the first file
from the given sequence that exists on it.
}

