% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project.init.R
\name{go}
\alias{go}
\title{Make a secret alias function so I don't have to type so much}
\usage{
go(codeDir = NULL, dataDir = NULL, RESOURCES = Sys.getenv("RESOURCES"))
}
\description{
Make a secret alias function so I don't have to type so much
}

