% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.config.R
\name{load.config}
\alias{load.config}
\title{Loads a yaml config file}
\usage{
load.config(project = NULL, sp = NULL, filename = NULL,
  usesPathsSection = FALSE)
}
\arguments{
\item{project}{A project (use default config file names for this project)}

\item{sp}{Subproject to activate}

\item{file}{file path to config file, allows you to specify an exact file.}
}
\description{
Loads a yaml config file
}
