% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{loadConfig}
\alias{loadConfig}
\title{Loads a yaml config file}
\usage{
loadConfig(project = NULL, sp = NULL, filename = NULL,
  usesPathsSectio = FALSE)
}
\arguments{
\item{project}{A project (use default config file names for this project)}

\item{sp}{Subproject to activate}

\item{file}{file path to config file, allows you to specify an exact file.}
}
\description{
Loads a yaml config file
}
