% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystemUtils.R
\name{makeAbsPath}
\alias{makeAbsPath}
\title{Create an absolute path from a primary target and a parent candidate.}
\usage{
makeAbsPath(perhapsRelative, parent)
}
\arguments{
\item{perhapsRelative:}{Path to primary target directory.}

\item{parent:}{Path to parent folder to use if target isn't absolute.}
}
\description{
Create an absolute path from a primary target and a parent candidate.
}
