% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{MakePath}
\alias{MakePath}
\title{Determine whether a path is absolute.}
\usage{
MakePath(target, envVar, whenNull)
}
\arguments{
\item{target}{The path to check for seeming absolute-ness.}

\item{env_var}{Name of the environment variable with parent folder candidate.}

\item{when_null}{Strategy for deriving target if its argument is null.}
}
\value{
\code{target} if it's already absolute, result of \code{when_null()} 
  if \code{target} is null, or joined version of parent candidate stored in 
  \code{env_var} and (relative) \code{target}.
}
\description{
Determine whether a path is absolute.
}
\seealso{
Other path operations: \code{\link{IsAbsolute}}
}
