% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.init.R
\docType{package}
\name{project.init}
\alias{project.init}
\alias{project.init-package}
\title{Package docs}
\usage{
project.init(codeDir = NULL, dataDir = NULL,
  RESOURCES = Sys.getenv("RESOURCES"))
}
\description{
Package docs

Define project init function
All this does is source the 00-init.R script for the project,
you pass a complete folder or a relative path.
}
\author{
Nathan Sheffield
}
\references{
\url{http://github.com/sheffien}
}

