% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{project.init2}
\alias{project.init2}
\title{Alias for backward compatibility}
\usage{
project.init2(codeDir = NULL, dataDir = NULL, subDir = NULL,
  resources = Sys.getenv("RESOURCES"))
}
\description{
Alias for backward compatibility
}
