% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.init.R
\name{project.refresh}
\alias{project.refresh}
\title{Helper alias to re-run init script, using your current dir settings.}
\usage{
project.refresh()
}
\description{
Helper alias to re-run init script, using your current dir settings.
}

