% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\docType{package}
\name{projectInit}
\alias{projectInit}
\alias{projectInit-package}
\title{Package docs}
\usage{
projectInit(projectName, codeDir = NULL, procDir = NULL,
  rawDir = NULL, webDir = NULL, outDir = NULL, resourcesDir = NULL,
  outputSubdir = NULL, scriptSubdir = "src", projectConfig = NULL,
  subproject = NULL)
}
\arguments{
\item{procDir}{Path to folder containing processed project data.}

\item{rawDir}{Path to folder containing raw project data.}

\item{outputSubdir}{Location for project-specific output, resolved by 
\code{dirOut} and stored as option \code{ROUT.SUBDIR}.}

\item{scriptSubdir}{Name for the folder within \code{codeRoot} that 
stores the scripts for this project.}

\item{subproject}{name of the subproject to be activated}

\item{codeRoot}{Path to the folder representing a code repository root.}

\item{resourcesRoot}{Location of general-purpose resourcesRoot; default is to use 
system environment variable \code{RESOURCES}.}
}
\description{
\code{projectInit} sources the \code{00-init.R} or \code{projectInit.R} 
script for the project. You pass a complete folder or a relative path.
}
\references{
\url{http://github.com/databio/projectInit}
}
\author{
Nathan Sheffield
}
