% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{projectLoad}
\alias{projectLoad}
\title{Loads a project without the rest of the init tasks.}
\usage{
projectLoad(projectName, pepConfig = NULL, subproject = NULL, loadrName = NULL)
}
\arguments{
\item{projectName}{A string identifying your project.}

\item{pepConfig}{Use to specify the (relative) location of your actual
PEP config file.}

\item{subproject}{name of the subproject to be activated}

\item{loadrName}{name to use in the shared variable environment. If not specified, 
will use projectName_subproject.}
}
\description{
Loads a project without the rest of the init tasks.
}
