% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{projectRefresh}
\alias{projectRefresh}
\title{Helper alias to re-run init script, using your current dir settings.}
\usage{
projectRefresh()
}
\description{
Helper alias to re-run init script, using your current dir settings.
}
