% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/packages.R
\name{refreshPackage}
\alias{refreshPackage}
\title{Detach a custom packages, re-document, re-install, and re-load.
Useful if I'm debugging packages and want to try the new version.
Expects it to be in the ~/rpack/ package folder. by default}
\usage{
refreshPackage(x, path = "~/rpack/", compileAttributes = TRUE,
  roxygenize = TRUE)
}
\arguments{
\item{x}{a package name}

\item{roxygenize}{should I roxygenize it, too?}
}
\description{
Detach a custom packages, re-document, re-install, and re-load.
Useful if I'm debugging packages and want to try the new version.
Expects it to be in the ~/rpack/ package folder. by default
}

