% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.init.R
\name{resourcedir}
\alias{resourcedir}
\title{Data Dir
Helper wrapper to get data for this project.}
\usage{
resourcedir(...)
}
\description{
Data Dir
Helper wrapper to get data for this project.
}

