% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-utils.R
\name{SetOption}
\alias{SetOption}
\title{Sets an option value if it's not already set.}
\usage{
SetOption(option, value, force = TRUE)
}
\arguments{
\item{option}{Name of option}

\item{value}{Value to set it to}

\item{force}{overwrite if already set.}
}
\description{
Sets an option value if it's not already set.
}
