% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{setOutputSubdir}
\alias{setOutputSubdir}
\title{Creates and sets outputSubdir}
\usage{
setOutputSubdir(...)
}
\arguments{
\item{...}{Arguments passed to \code{ffProc()}.}
}
\description{
Creates and sets outputSubdir
}
