% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{toggleError}
\alias{toggleError}
\title{Quick shortcut procedure for toggling error mode (for my convenience)
te for Toggle Error}
\usage{
toggleError()
}
\description{
Quick shortcut procedure for toggling error mode (for my convenience)
te for Toggle Error
}
