% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.init.R, R/random-utils.R
\name{utility}
\alias{utility}
\alias{utility}
\title{This is Nathan's custom utility loading function.
It goes with my R utility package, and is used to load up utilities from
that package. It was originally in funcCommon.R, but I moved it here
because that's actually a utility, which uses this function to load it,
so it's more natural if this loading function is here.}
\usage{
utility(utilities, utilityDir = "")

utility(utilities, utilityDir = "")
}
\arguments{
\item{utilityDir}{Directory to search (custom)}

\item{utility}{The name of the R script in the util directory to load.}

\item{utility}{The name of the R script in the util directory to load.}

\item{utilityDir}{Directory to search (custom)}
}
\description{
It will search first any passed dir, then the working dir, and finally
the global RGENOMEUTILS for the script to load.

It will search first any passed dir, then the working dir, and finally
the global RGENOMEUTILS for the script to load.
}
