% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.init.R
\name{createOutputSubdir}
\alias{createOutputSubdir}
\title{Helper function to silently create a subdirectory in the project
output directory.}
\usage{
createOutputSubdir(...)
}
\description{
Helper function to silently create a subdirectory in the project
output directory.
}

