% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_aliases.R, R/project.init.R
\name{createRootSubdir}
\alias{createRootSubdir}
\alias{createRootSubdir}
\title{Helper function to silently create a subdirectory in the parent project
directory (the processed data directory).}
\usage{
createRootSubdir(...)

createRootSubdir(...)
}
\description{
Helper function to silently create a subdirectory in the parent project
directory (the processed data directory).

Helper function to silently create a subdirectory in the parent project
directory (the processed data directory).
}
