% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project.init.R
\name{dirout}
\alias{dirout}
\title{ROUT Wrapper function}
\usage{
dirout(...)
}
\description{
pass a relative file path, and this will append the global results
directory to it. Use it to stick output easily directly into the results
directory, instead of relative to the local directory.
This allows you to keep a working directory that's relative to your code,
but put the results somewhere else (which is shared space).
}

