% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{go}
\alias{go}
\title{Make a secret alias function so I don't have to type so much}
\usage{
go(codeRoot = NULL, dataDir = NULL, rawDir = NULL,
  outputSubdir = NULL, resources = Sys.getenv("RESOURCES"),
  scriptSubdir = "src", subproject = NULL)
}
\description{
Make a secret alias function so I don't have to type so much
}
