% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{go}
\alias{go}
\title{Make a secret alias function so I don't have to type so much}
\usage{
go(codeDir = NULL, dataDir = NULL, subDir = NULL,
  resources = Sys.getenv("RESOURCES"))
}
\description{
Make a secret alias function so I don't have to type so much
}
