% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{go}
\alias{go}
\title{Alias so I don't have to type so much}
\usage{
go(projectName, codeDir = NULL, procDir = NULL, rawDir = NULL,
  webDir = NULL, outDir = NULL, resourcesDir = NULL,
  outputSubdir = NULL, scriptSubdir = "src", projectConfig = NULL,
  subproject = NULL)
}
\description{
Alias so I don't have to type so much
}
