% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.init.R
\name{go}
\alias{go}
\title{Make a secret alias function so I don't have to type so much}
\usage{
go(code_dir = NULL, data_dir = NULL, RESOURCES = Sys.getenv("RESOURCES"))
}
\description{
Make a secret alias function so I don't have to type so much
}
