% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project.init.R
\name{project.init}
\alias{project.init}
\title{Define project init function
All this does is source the 00-init.R script for the project,
you pass a complete folder or a relative path.}
\usage{
project.init(codeDir = NULL, dataDir = NULL,
  RESOURCES = Sys.getenv("RESOURCES"))
}
\description{
Define project init function
All this does is source the 00-init.R script for the project,
you pass a complete folder or a relative path.
}

