% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{project.init}
\alias{project.init}
\title{Alias for backward compatibility}
\usage{
project.init(codeRoot = NULL, dataDir = NULL, outputSubdir = NULL,
  resources = Sys.getenv("RESOURCES"), scriptSubdir = "src")
}
\description{
Alias for backward compatibility
}
