% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\docType{package}
\name{project.init}
\alias{project.init}
\alias{project.init-package}
\alias{project.init}
\title{Package docs}
\usage{
project.init(codeDir = NULL, dataDir = NULL, subDir = NULL,
  resources = Sys.getenv("RESOURCES"))
}
\description{
Package docs

Alias for backward compatibility
}
\references{
\url{http://github.com/databio/project.init}
}
\author{
Nathan Sheffield
}
