% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{project.init2}
\alias{project.init2}
\title{Alias for backward compatibility}
\usage{
project.init2(codeRoot = NULL, dataDir = NULL, rawDir = NULL,
  outputSubdir = NULL, resources = Sys.getenv("RESOURCES"),
  scriptSubdir = "src")
}
\description{
Alias for backward compatibility
}
