% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.init.R
\name{project.init2}
\alias{project.init2}
\title{Helper alias for the common case where the data and code dirs share
a name.}
\usage{
project.init2(codeDir)
}
\description{
Helper alias for the common case where the data and code dirs share
a name.
}

