% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{project.init2}
\alias{project.init2}
\title{Alias for backward compatibility}
\usage{
project.init2(projectName, codeDir = NULL, procDir = NULL,
  rawDir = NULL, webDir = NULL, outDir = NULL, resourcesDir = NULL,
  outputSubdir = NULL, scriptSubdir = "src", projectConfig = NULL,
  subproject = NULL)
}
\description{
Alias for backward compatibility
}
