% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{projectInit}
\alias{projectInit}
\title{Initialize workspace for a given project}
\usage{
projectInit(codeDir = NULL, dataDir = NULL, subDir = NULL,
  resources = Sys.getenv("RESOURCES"))
}
\arguments{
\item{codeDir}{}
}
\description{
This function will source the 00-init.R script for the project.
You pass a complete folder or a relative path.
}
