% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\name{projectInit}
\alias{projectInit}
\title{Project workspace initializer}
\usage{
projectInit(codeRoot = NULL, dataDir = NULL, outputSubdir = NULL,
  resources = Sys.getenv("RESOURCES"), scriptSubdir = "src")
}
\arguments{
\item{codeRoot}{Path to the folder representing a code repository root.}

\item{dataDir}{Path to folder containing project data.}

\item{outputSubdir}{Location for project-specific output, resolved by 
\code{dirOut} and stored as option \code{ROUT.SUBDIR}.}

\item{resources}{Location of general-purpose resources; default is to use 
system environment variable \code{RESOURCES}.}

\item{scriptSubdir}{Name for the folder within \code{codeRoot} that 
stores the scripts for this project.}
}
\description{
\code{projectInit} sources the \code{00-init.R} or \code{projectInit.R} 
script for the project. You pass a complete folder or a relative path.
}
