% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectInit.R
\docType{package}
\name{projectInit}
\alias{projectInit}
\alias{projectInit-package}
\title{Package docs}
\usage{
projectInit(codeRoot = NULL, dataDir = NULL, rawDir = NULL,
  outputSubdir = NULL, resources = Sys.getenv("RESOURCES"),
  scriptSubdir = "src", subproject = NULL)
}
\arguments{
\item{codeRoot}{Path to the folder representing a code repository root.}

\item{dataDir}{Path to folder containing processed project data.}

\item{rawDir}{Path to folder containing raw project data.}

\item{outputSubdir}{Location for project-specific output, resolved by 
\code{dirOut} and stored as option \code{ROUT.SUBDIR}.}

\item{resources}{Location of general-purpose resources; default is to use 
system environment variable \code{RESOURCES}.}

\item{scriptSubdir}{Name for the folder within \code{codeRoot} that 
stores the scripts for this project.}

\item{subproject}{name of the subproject to be activated}
}
\description{
\code{projectInit} sources the \code{00-init.R} or \code{projectInit.R} 
script for the project. You pass a complete folder or a relative path.
}
\references{
\url{http://github.com/databio/projectInit}
}
\author{
Nathan Sheffield
}
