% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project.init.R
\name{setOption}
\alias{setOption}
\title{Sets an option value if it's not already set.}
\usage{
setOption(option, value)
}
\arguments{
\item{option}{Name of option}

\item{value}{Value to set it to}
}
\description{
Sets an option value if it's not already set.
}

