% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project.init.R
\name{utility}
\alias{utility}
\title{This is Nathan's custom utility loading function.
It goes with my R utility package, and is used to load up utilities from
that package. It was originally in funcCommon.R, but I moved it here
because that's actually a utility, which uses this function to load it,
so it's more natural if this loading function is here.}
\usage{
utility(utilities, utilityDir = "")
}
\arguments{
\item{utilityDir}{Directory to search (custom)}

\item{utility}{The name of the R script in the util directory to load.}
}
\description{
It will search first any passed dir, then the working dir, and finally
the global SHARE.RUTIL.DIR for the script to load.
}
\details{
Should probably be renamed in the future.
}

